backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
pandas==2.1.3
Pillow==10.1.0
django-filter==23.4
drf-spectacular==0.26.5
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-redis==5.4.0 